feat:advisories Consolidate the types into utils and delete the unused banner component.
