--- sites/main-site/src/components/advisory/AdvisoryBanner.astro
@@ -1,62 +1,0 @@
----
-import AdvisoryBannerElement from "@components/advisories/AdvisoryBannerElement.svelte";
-import { getCollection } from "astro:content";
-import type { CollectionEntry } from "astro:content";
-import { markdown } from "@astropub/md";
-import { advisories_type_classes, advisories_type_icons } from "./advisoryTypes";
-
-let advisories: CollectionEntry<"advisories">[] = await getCollection("advisories");
-advisories = await Promise.all(
-    advisories.map(async (advisories) => {
-        advisories.data.subtitle = `${await markdown(advisories.data.subtitle)}`;
-        return advisories;
-    }),
-);
-let now = new Date().getTime();
-advisories = advisories.sort((a, b) => {
-    if (!a.data.publishedDate || !b.data.publishedDate) return 0;
-    return b.data.publishedDate.getTime() - a.data.publishedDate.getTime();
-});
-
-advisories = advisories.filter((advisories) => {
-    return advisories.id.split("/").length === 2;
-});
-const currentAdvisory = advisories.filter((advisories) => {
-    // Calculate date 3 months ago
-    const threeMonthsAgo = new Date();
-    threeMonthsAgo.setMonth(threeMonthsAgo.getMonth() - 3);
-    const threeMonthsAgoTime = threeMonthsAgo.getTime();
-
-    // Check if advisory was published within the last 3 months
-    const publishedDateUnix = advisories.data.publishedDate?.getTime();
-    return publishedDateUnix && publishedDateUnix >= threeMonthsAgoTime;
-});
-
-{
-    currentAdvisory.length > 0 && (
-        <div>
-            <div class="current-advisories-container bg-body-secondary">
-                <AdvisoryBannerElement
-                    advisories_time_category={"ongoing"}
-                    advisories={currentAdvisory}
-                    {advisories_type_classes}
-                    {advisories_type_icons}
-                    client:load
-                />
-            </div>
-            <div class="triangle-down bg-body-secondary d-none " />
-        </div>
-    )
-}
-<style lang="scss">
-    @import "@styles/_variables.scss";
-    .current-advisories-container:has(.ongoing-advisories) ~ .triangle-down {
-        display: block !important;
-    }
-
-    @include media-breakpoint-down(md) {
-        .mainpage-subheader-heading {
-            padding-top: 2rem;
-        }
-    }
-</style>
