+++ sites/main-site/src/components/advisory/AdvisoryHeader.astro
@@ -1,6 +1,6 @@
 ---
 import LocalDateTime from "@components/event/LocalDateTime.svelte";
-import { advisories_type_classes, advisories_type_icons } from "./advisoryTypes";
+import { advisories_type_classes, advisories_type_icons } from "./advisoryUtils";
 import GitHubProfilePictureExtended from "@components/GitHubProfilePictureExtended.svelte";
 import {
     formatAdvisoryType,
